fix schema change
